requirements.txt
Flask==3.0.3
flask-cors>=4.0
mysql-connector-python==8.2.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.1
# Optional for mood prediction module
tensorflow==2.16.1
